Converted ball position/velocity to doubles and made them independant of the update rate.  Also added many comments.

+++ src/bam/pong/Ball.java
@@ -7,17 +7,21 @@ package bam.pong;
  * @author Brian
  */
 public class Ball {
-	int x, y;
-	int dx, dy;
+	
+	/** Current location of the ball. */
+	double x, y;
+	
+	/** Movement vector in pixels/second. */
+	double dx, dy;
 
-	public Ball(int x, int y, int dx, int dy) {
+	public Ball(double x, double y, double dx, double dy) {
 		this.x = x;
 		this.y = y;
 		this.dx = dx;
 		this.dy = dy;
 	}
 	
-	public Ball(int x, int y) {
+	public Ball(double x, double y) {
 		this(x, y, 0, 0);
 	}
